% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_confdiv}
\alias{check_confdiv}
\title{Checks if division or conference is valid}
\usage{
check_confdiv(group = NULL, value = NULL, teams = NULL)
}
\arguments{
\item{group}{Character string for group ("div" or "conf").}

\item{value}{Character string for group's value (e.g., 1 or "Big Ten")}
}
\description{
Checks if division or conference is valid
}
\keyword{internal}
