% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_contest}
\alias{check_contest}
\title{Checks if contest ID is valid}
\usage{
check_contest(contest = NULL)
}
\arguments{
\item{contest}{Contest ID}
}
\description{
Checks if contest ID is valid
}
\keyword{internal}
