% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_match}
\alias{check_match}
\title{Checks if value is matched in vector}
\usage{
check_match(name = NULL, value = NULL, vec = NULL)
}
\arguments{
\item{name}{Argument name.}

\item{value}{Value.}

\item{vec}{Vector.}
}
\description{
Checks if value is matched in vector
}
\keyword{internal}
