% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_sport}
\alias{check_sport}
\title{Checks if sport is valid}
\usage{
check_sport(sport, vb_only = TRUE)
}
\arguments{
\item{sport}{Sport code.}

\item{vb_only}{Logical indicating whether to check only for
volleyall sports (TRUE) or all sports (FALSE)}
}
\description{
Checks if sport is valid
}
\keyword{internal}
