% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_team_id}
\alias{check_team_id}
\title{Checks if team ID is valid}
\usage{
check_team_id(team_id = NULL)
}
\arguments{
\item{team_id}{Team ID}
}
\description{
Checks if team ID is valid
}
\keyword{internal}
