% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_team_name}
\alias{check_team_name}
\title{Checks if team name is valid}
\usage{
check_team_name(team = NULL, teams = NULL)
}
\arguments{
\item{team}{Team name}

\item{teams}{Data frame of team names}
}
\description{
Checks if team name is valid
}
\keyword{internal}
