% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_year}
\alias{check_year}
\title{Checks if year is valid}
\usage{
check_year(year = NULL, single = FALSE)
}
\arguments{
\item{year}{Year.}

\item{single}{Logical for whether year should be a single element or can be
a vector of multiple years.}
}
\description{
Checks if year is valid
}
\keyword{internal}
