% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conference_match_stats.R
\name{conference_match_stats}
\alias{conference_match_stats}
\title{Aggregate player match statistics for a NCAA conference and seasons}
\usage{
conference_match_stats(
  year = NULL,
  conf = NULL,
  sport = "WVB",
  save = FALSE,
  path = "."
)
}
\arguments{
\item{year}{Year for fall of desired season.}

\item{conf}{NCAA conference name.}

\item{sport}{Three letter abbreviation for NCAA sport (must be upper case;
for example "WVB" for women's volleyball and "MVB" for men's volleyball).}

\item{save}{Logical for whether to save the statistics locally as CSVs
(default FALSE).}

\item{path}{Character string of path to save statistics files.}
}
\value{
Returns data frame of player match statistics.
}
\description{
This is a wrapper around \code{\link[=group_stats]{group_stats()}} that extracts match data from
players in all teams in the chosen conference.
Conferences names can be found in
\link{ncaa_conferences}.
}
\examples{
\dontrun{
conference_match_stats(year = 2024, conf = "Big Ten")
conference_match_stats(year = 2023, conf = "Big West", sport = "MVB")
conference_match_stats(year = 2024, conf = "Big Ten", save = TRUE, path = "data/")
}
}
\seealso{
Other functions that aggregate statistics: 
\code{\link{conference_pbp_stats}()},
\code{\link{conference_season_stats}()},
\code{\link{division_match_stats}()},
\code{\link{division_pbp_stats}()},
\code{\link{division_season_stats}()},
\code{\link{group_stats}()}
}
\concept{functions that aggregate statistics}
