% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conference_season_stats.R
\name{conference_season_stats}
\alias{conference_season_stats}
\title{Aggregate player statistics from a particular conference and season}
\usage{
conference_season_stats(
  year = NULL,
  conf = NULL,
  sport = "WVB",
  save = FALSE,
  path = ""
)
}
\arguments{
\item{year}{Year for fall of desired season.}

\item{conf}{NCAA conference name.}

\item{sport}{Three letter abbreviation for NCAA sport (must be upper case;
for example "WVB" for women's volleyball and "MVB" for men's volleyball).}

\item{save}{Logical for whether to save the statistics locally as CSVs
(default FALSE).}

\item{path}{Character string of path to save statistics files.}
}
\value{
Returns list with data frames of player statistics and team statistics.
}
\description{
This is a wrapper around \code{\link[=group_stats]{group_stats()}} that extracts all season data from
players in all teams from the chosen conference.
It aggregates all player data and team data into separate data frames and
combines them into a list. Conferences names can be found in
\link{ncaa_conferences}.
}
\examples{
\dontrun{
conference_season_stats(year = 2024, conf = "Big Ten")
conference_season_stats(year = 2023, conf = "Big West", sport = "MVB")
conference_season_stats(year = 2024, conf = "Big Ten", save = TRUE, path = "data/")
}
}
\seealso{
Other functions that aggregate statistics: 
\code{\link{division_season_stats}()},
\code{\link{group_stats}()}
}
\concept{functions that aggregate statistics}
