% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/division_season_stats.R
\name{division_player_stats}
\alias{division_player_stats}
\title{Aggregate player statistics from a particular division and season}
\usage{
division_player_stats(...)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{division_player_stats()} was renamed to \code{division_season_stats()} to
create a more consistent API.
}
\keyword{internal}
