% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_chromium}
\alias{find_chromium}
\title{Find path for Chrome or Chromium}
\usage{
find_chromium()
}
\description{
Uses \code{chromote::find_chrome()} to find path of Google Chrome or Chromium
}
\keyword{internal}
