% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_team_info}
\alias{get_team_info}
\title{Gets year, team, and conference from team ID}
\usage{
get_team_info(team_id = NULL)
}
\arguments{
\item{team_id}{Team ID}
}
\description{
Gets year, team, and conference from team ID
}
\keyword{internal}
