% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_player_stats.R
\name{group_player_stats}
\alias{group_player_stats}
\title{Aggregate all player statistics from a vector of teams in a year}
\usage{
group_player_stats(teams, year, sport = "WVB")
}
\arguments{
\item{teams}{Character vector of team names to aggregate.}

\item{year}{Year for fall volleyball season.}

\item{sport}{Three letter abbreviation for NCAA sport (must be upper case;
for example "WVB" for women's volleyball and "MVB" for men's volleyball).}
}
\value{
Returns list with player statistics and team statistics data frames.
}
\description{
Aggregate all player statistics from a vector of teams in a year
}
\examples{
\dontrun{
group_player_stats(c("Nebraska", "Wisconsin"), 2024)
group_player_stats(c("UCLA", "Long Beach St."), 2023, sport = "MVB")
}
}
\seealso{
Other aggregate statistics: 
\code{\link{conference_player_stats}()},
\code{\link{division_player_stats}()}
}
\concept{aggregate statistics}
