% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{html_table_raw}
\alias{html_table_raw}
\title{Creates table of raw HTML}
\usage{
html_table_raw(
  x,
  header = NA,
  trim = TRUE,
  dec = ".",
  na.strings = "NA",
  convert = TRUE
)
}
\description{
Copied and modified from \code{{rvest}}
https://github.com/tidyverse/rvest/blob/main/R/table.R
}
\keyword{internal}
