% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ncaa_conferences}
\alias{ncaa_conferences}
\title{NCAA Conference Names}
\format{
A character vector with 111 conference names.
}
\source{
\url{https://stats.ncaa.org}
}
\usage{
ncaa_conferences
}
\description{
This vector includes names for all NCAA volleyball conferences.
}
\examples{
head(ncaa_conferences)
}
\seealso{
Other data sets: 
\code{\link{mvb_teams}},
\code{\link{ncaa_sports}},
\code{\link{ncaa_teams}},
\code{\link{wvb_teams}}
}
\concept{data sets}
\keyword{datasets}
