% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ncaa_sports}
\alias{ncaa_sports}
\title{NCAA Sports and Sport Codes}
\format{
A data frame with 100 rows and 2 columns:
\describe{
\item{code}{Sport code}
\item{sport}{Sport name}
}
}
\source{
\url{https://ncaaorg.s3.amazonaws.com/championships/resources/common/NCAA_SportCodes.pdf}
}
\usage{
ncaa_sports
}
\description{
This data frame includes all NCAA women's and men's sports and the codes
used to refer to the sports.
}
\examples{
head(ncaa_sports)
}
\seealso{
Other data sets: 
\code{\link{mvb_teams}},
\code{\link{ncaa_conferences}},
\code{\link{ncaa_teams}},
\code{\link{wvb_teams}}
}
\concept{data sets}
\keyword{datasets}
