% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ncaa_teams}
\alias{ncaa_teams}
\title{NCAA Team Names}
\format{
A character vector with 1,089 team names.
}
\source{
\url{https://stats.ncaa.org}
}
\usage{
ncaa_teams
}
\description{
This vector includes names for all NCAA volleyball teams.
}
\examples{
head(ncaa_teams)
}
\seealso{
Other data sets: 
\code{\link{mvb_teams}},
\code{\link{ncaa_conferences}},
\code{\link{ncaa_sports}},
\code{\link{wvb_teams}}
}
\concept{data sets}
\keyword{datasets}
