% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncaa_wvolleyball_boxscore.R
\name{ncaa_wvolleyball_boxscore}
\alias{ncaa_wvolleyball_boxscore}
\title{Get NCAA volleyball team boxscore from a season}
\usage{
ncaa_wvolleyball_boxscore(team, year)
}
\arguments{
\item{team_name}{team name, found in \code{\link[=ncaa_wvolleyball_teams]{ncaa_wvolleyball_teams()}}}

\item{yr}{year YYYY format}
}
\value{
ncaa_wvolleyball_boxscore data.frame: 'Player', 'Pos', 'S', 'MP', 'Kills', 'Errors', 'Total Attacks', 'Hit Pct', 'Assists', 'Aces', 'SErr', 'Digs', 'RErr', 'Block Solos', 'Block Assists', 'BErr', 'PTS', 'BHE', 'team', 'opponent', 'date', 'game_id', 'away', 'home', 'away_sets', 'home_sets', 'away_points', 'away_pointpct', 'home_pointpct'
}
\description{
Get NCAA volleyball team boxscore from a season
}
\examples{
try(ncaa_wvolleyball_boxscore())

\dontrun{
  ncaa_wvolleyball_boxscore('Yale', 2019)
  }

}
\author{
Tyler Widdison
}
