% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncaa_volleyball_conf_schedule.R
\name{ncaa_wvolleyball_conf_schedule}
\alias{ncaa_wvolleyball_conf_schedule}
\title{Get NCAA volleyball conference schedule}
\usage{
ncaa_wvolleyball_conf_schedule(conf_name, yr)
}
\arguments{
\item{conf_name}{conference name, found in ncaa_wvolleyball_teams()}

\item{yr}{year YYYY format}
}
\value{
ncaa_wvolleyball_conf_schedule data.frame: 'url', 'date', 'home', 'away', 'location'
}
\description{
Get NCAA volleyball conference schedule
}
\examples{
try(ncaa_wvolleyball_conf_schedule())

  ncaa_wvolleyball_conf_schedule('Big Ten', 2023)

}
\author{
Tyler Widdison
}
