% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncaa_wvolleyball_pbp.R
\name{ncaa_wvolleyball_pbp}
\alias{ncaa_wvolleyball_pbp}
\title{Get NCAA volleyball team play by play from a season}
\usage{
ncaa_wvolleyball_pbp(team, year)
}
\arguments{
\item{team_name}{team name, found in \code{\link[=ncaa_wvolleyball_teams]{ncaa_wvolleyball_teams()}}}

\item{yr}{year YYYY format}
}
\value{
ncaa_wvolleyball_pbp data.frame: 'set_number', 'skill', 'team', 'player_name', 'away_score', 'home_score', 'serving_team', 'away_team', 'home_team', 'opponent', 'point_won_by', 'ncaa_match_id', 'point_id', 'id', 'location', 'box_score', 'play_by_play'
}
\description{
Return a play by play dataframe
}
\examples{
try(ncaa_wvolleyball_pbp())

\dontrun{
  ncaa_wvolleyball_pbp('Yale', 2019)
  }

}
\author{
Tyler Widdison
}
