% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncaa_volleyball_schedule.R
\name{ncaa_wvolleyball_schedule}
\alias{ncaa_wvolleyball_schedule}
\title{Get NCAA volleyball team schedule}
\usage{
ncaa_wvolleyball_schedule(team_name, team_id, yr)
}
\arguments{
\item{team_name}{team name, found in \code{\link[=ncaa_wvolleyball_teams]{ncaa_wvolleyball_teams()}}}

\item{team_id}{team id, found in \code{\link[=ncaa_wvolleyball_teams]{ncaa_wvolleyball_teams()}}}

\item{yr}{year YYYY format}
}
\value{
ncaa_wvolleyball_schedule data.frame: 'url', 'date', 'home', 'away', 'location'
}
\description{
Get NCAA volleyball team schedule
}
\examples{
try(ncaa_wvolleyball_schedule())

\dontrun{
  ncaa_wvolleyball_schedule('Yale', 813, 2019)
  }

}
\author{
Tyler Widdison
}
