% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncaa_volleyball_teams.R
\name{ncaa_wvolleyball_teams}
\alias{ncaa_wvolleyball_teams}
\title{Get all NCAA volleyball team names and ids}
\usage{
ncaa_wvolleyball_teams()
}
\value{
ncaa_wvolleyball_teams data.frame: 'Name', 'Conference', 'ID', NOTE: Only Division 1 conferences are returned
}
\description{
Get all NCAA volleyball team names and ids
}
\examples{
try(ncaa_wvolleyball_teams())

\dontrun{
  get_ncaaw_teams()
  }

}
\author{
Tyler Widdison
}
