% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncaavolleyballr-package.R
\docType{package}
\name{ncaavolleyballr-package}
\alias{ncaavolleyballr}
\alias{ncaavolleyballr-package}
\title{ncaavolleyballr: Extract Data from NCAA Women's and Men's Volleyball Website}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Extracts team records/schedules and player statistics for the 2020-2024 National Collegiate Athletic Association (NCAA) women's and men's divisions I, II, and III volleyball teams from \url{https://stats.ncaa.org}. Functions can aggregate statistics for teams, conferences, divisions, or custom groups of teams.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/JeffreyRStevens/ncaavolleyballr}
  \item \url{https://jeffreyrstevens.github.io/ncaavolleyballr/}
  \item Report bugs at \url{https://github.com/JeffreyRStevens/ncaavolleyballr/issues}
}

}
\author{
\strong{Maintainer}: Jeffrey R. Stevens \email{jeffrey.r.stevens@protonmail.com} (\href{https://orcid.org/0000-0003-2375-1360}{ORCID}) [copyright holder]

}
\keyword{internal}
