% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player_match_stats.R
\name{player_match_stats}
\alias{player_match_stats}
\title{Extract player statistics for a particular match}
\usage{
player_match_stats(
  contest = NULL,
  team = NULL,
  team_stats = TRUE,
  sport = "WVB"
)
}
\arguments{
\item{contest}{Contest ID determined by NCAA for match. To find ID, use
\code{\link[=find_team_contests]{find_team_contests()}} for a team and season.}

\item{team}{Name of school. Must match name used by NCAA. Find exact team
name with \code{\link[=find_team_name]{find_team_name()}}.}

\item{team_stats}{Logical indicating whether to include (TRUE) or exclude
(FALSE) team statistics. Default includes team statistics with player
statistics.}

\item{sport}{Three letter abbreviation for NCAA sport (must be upper case;
for example "WVB" for women's volleyball and "MVB" for men's volleyball).}
}
\value{
By default, returns list with two data frames: home and away team match
statistics. If team is specified, only single data frame is returned.
}
\description{
The NCAA's page for a match/contest includes a tab called
"Individual Statistics". This function extracts the tables of player
match statistics for both home and away teams, as well as team statistics
(though these can be omitted). If a particular team is specified, only that
team's statistics will be returned.
}
\examples{
\dontrun{
player_match_stats(contest = "6080706")
player_match_stats(contest = "6080706", team = "Louisville")
player_match_stats(contest = "4475421", team_stats = FALSE, sport = "MVB")
}
}
\seealso{
Other functions that extract player statistics: 
\code{\link{player_season_stats}()}
}
\concept{functions that extract player statistics}
