% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player_stats.R
\name{player_stats}
\alias{player_stats}
\title{Extract player season statistics}
\usage{
player_stats(team_id)
}
\arguments{
\item{team_id}{Team ID determined by NCAA for season. To find ID, use
[`find_team_id()`].}
}
\value{
Tibble of player statistics.
}
\description{
Extract player season statistics
}
\examples{
\dontrun{
player_stats("585290")
}
}
