% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{request_url}
\alias{request_url}
\title{Submit URL request, check, and return response}
\usage{
request_url(url)
}
\arguments{
\item{url}{URL for request.}
}
\description{
Submit URL request, check, and return response
}
\note{
This function \strong{requires internet connectivity} as it checks the
\href{https://stats.ncaa.org}{NCAA website} for information.
}
\keyword{internal}
