% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{table_fill}
\alias{table_fill}
\title{Table filling algorithm}
\usage{
table_fill(cells, trim = TRUE)
}
\description{
Copied and modified from \code{{rvest}}
https://github.com/tidyverse/rvest/blob/main/R/table.R
}
\keyword{internal}
