% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_match_stats.R
\name{team_match_stats}
\alias{team_match_stats}
\title{Extract team summary statistics for all matches in a particular season}
\usage{
team_match_stats(team_id = NULL, opponent = FALSE, sport = "WVB")
}
\arguments{
\item{team_id}{Team ID determined by NCAA for season. To find ID, use
\code{\link[=find_team_id]{find_team_id()}}.}

\item{opponent}{Logical indicating whether to include team's stats
(FALSE) or opponent's stats (TRUE). Default is set to FALSE, returning
team stats.}

\item{sport}{Three letter abbreviation for NCAA sport (must be upper case;
for example "WVB" for women's volleyball and "MVB" for men's volleyball).}
}
\value{
Returns a data frame of summary team statistics for each match of the season.
}
\description{
The NCAA's main page for a team includes a tab called "Game By Game"
and a section called "Game by Game Stats".
This function extracts the team's summary statistics for each match of the
season.
}
\examples{
\donttest{
team_match_stats(team_id = "585290")
}
}
\seealso{
Other functions that extract team statistics: 
\code{\link{team_season_info}()},
\code{\link{team_season_stats}()}
}
\concept{functions that extract team statistics}
