% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player_season_stats.R
\name{team_player_stats}
\alias{team_player_stats}
\title{Extract player statistics from a particular team and season}
\usage{
team_player_stats(...)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{team_player_stats()} was renamed to \code{player_season_stats()} to create a more
consistent API.
}
\keyword{internal}
