% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_season.R
\name{team_season}
\alias{team_season}
\title{Extract arena, coach, record, and schedule information for a season}
\usage{
team_season(team_id)
}
\arguments{
\item{team_id}{Team ID determined by NCAA for season. To find ID, use
[`find_team_id()`].}
}
\value{
List that includes arena, coach, record, and schedule information.
}
\description{
Extract arena, coach, record, and schedule information for a season
}
\examples{
\dontrun{
team_season("585290")
}
}
