% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_season_stats.R
\name{team_season_stats}
\alias{team_season_stats}
\title{Extract arena, coach, record, and schedule information for a season}
\usage{
team_season_stats(team_id)
}
\arguments{
\item{team_id}{Team ID determined by NCAA for season. To find ID, use
`find_team_id()`.}
}
\value{
List that includes arena, coach, record, and schedule information.
}
\description{
Extract arena, coach, record, and schedule information for a season
}
\examples{
\dontrun{
team_season_stats("585290")
team_season_stats(find_team_id("Nebraska", 2024))
team_season_stats(find_team_id("UCLA", 2023, sport = "MVB"))
}
}
