% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volleyballr-package.R
\docType{package}
\name{volleyballr-package}
\alias{volleyballr}
\alias{volleyballr-package}
\title{volleyballr: Extract Data from NCAA Women's and Men's Volleyball Website}
\description{
Downloads NCAA Division 1, 2, and 3 women's and men's volleyball team information from 2020-2024, including records, schedules, and player statistics.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/JeffreyRStevens/volleyballr}
  \item Report bugs at \url{https://github.com/JeffreyRStevens/volleyballr/issues}
}

}
\author{
\strong{Maintainer}: Jeffrey R. Stevens \email{jeffrey.r.stevens@protonmail.com} (\href{https://orcid.org/0000-0003-2375-1360}{ORCID}) [copyright holder]

}
\keyword{internal}
