% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conference_player_stats.R
\name{conference_player_stats}
\alias{conference_player_stats}
\title{Aggregate player statistics from a particular conference and season}
\usage{
conference_player_stats(
  year = NULL,
  conf = NULL,
  sport = "WVB",
  save = FALSE,
  path = ""
)
}
\arguments{
\item{year}{Year for fall of desired season.}

\item{conf}{NCAA conference name.}

\item{sport}{Three letter abbreviation for NCAA sport (must be upper case;
for example "WVB" for women's volleyball and "MVB" for men's volleyball).}

\item{save}{Logical for whether to save the statistics locally as CSVs
(default FALSE).}

\item{path}{Character string of path to save statistics files.}
}
\value{
Returns list with player statistics and team statistics data frames.
}
\description{
This is a wrapper around \code{\link{group_player_stats}()} that extracts
all season data from players in all teams from the chosen conference.
It aggregates all player data and team data into separate data frames and
combines them into a list. Conferences names can be found in
\code{\link{ncaa_conferences}}.
}
\examples{
\dontrun{
conference_player_stats(2024, conf = "Big Ten")
conference_player_stats(2023, conf = "Big West", sport = "MVB")
conference_player_stats(2024, conf = "Big Ten", save = TRUE, path = "data/")
}
}
\seealso{
Other functions that aggregate statistics: 
\code{\link{division_player_stats}()},
\code{\link{group_player_stats}()}
}
\concept{functions that aggregate statistics}
