% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conference_player_stats.R
\name{conference_player_stats}
\alias{conference_player_stats}
\title{Aggregate all player statistics from a particular conference and year}
\usage{
conference_player_stats(
  year,
  conf = NULL,
  sport = "WVB",
  save = FALSE,
  path = ""
)
}
\arguments{
\item{year}{Year for fall volleyball season.}

\item{sport}{Three letter abbreviation for NCAA sport (must be upper case;
for example "WVB" for women's volleyball and "MVB" for men's volleyball).}

\item{save}{Logical for whether to save the statistics locally as CSVs
(default FALSE).}

\item{path}{Character string of path to save statistics files.}

\item{conference}{NCAA conference name.}
}
\value{
Returns list with player statistics and team statistics data frames.
}
\description{
Aggregate all player statistics from a particular conference and year
}
\examples{
\dontrun{
conference_player_stats(2024)
conference_player_stats(2023, division = 3, sport = "MVB")
conference_player_stats(2024, save = TRUE, path = "data/")
}
}
\seealso{
Other aggregate statistics: 
\code{\link{division_player_stats}()}
}
\concept{aggregate statistics}
