% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conference_season_stats.R
\name{conference_player_stats}
\alias{conference_player_stats}
\title{Aggregate player statistics from a particular conference and season}
\usage{
conference_player_stats(...)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{conference_player_stats()} was renamed to \code{conference_season_stats()} to
create a more consistent API.
}
\keyword{internal}
