% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/division_player_stats.R
\name{division_player_stats}
\alias{division_player_stats}
\title{Aggregate all player statistics from a particular division and year}
\usage{
division_player_stats(
  year,
  division = 1,
  sport = "WVB",
  save = FALSE,
  path = ""
)
}
\arguments{
\item{year}{Year for fall volleyball season.}

\item{division}{NCAA division (must be 1, 2, or 3 for women's and 1 or 3
for men's).}

\item{sport}{Three letter abbreviation for NCAA sport (must be upper case;
for example "WVB" for women's volleyball and "MVB" for men's volleyball).}

\item{save}{Logical for whether to save the statistics locally as CSVs
(default FALSE).}

\item{path}{Character string of path to save statistics files.}
}
\value{
Returns list with player statistics and team statistics data frames.
}
\description{
Aggregate all player statistics from a particular division and year
}
\examples{
\dontrun{
division_player_stats(2024)
division_player_stats(2023, division = 3, sport = "MVB")
division_player_stats(2024, save = TRUE, path = "data/")
}
}
\seealso{
Other aggregate statistics: 
\code{\link{conference_player_stats}()}
}
\concept{aggregate statistics}
