% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_team_id.R
\name{find_team_id}
\alias{find_team_id}
\title{Find team ID for year}
\usage{
find_team_id(name, year)
}
\arguments{
\item{name}{Name of school. Must match name used by NCAA.}

\item{year}{Year for fall volleyball season}
}
\value{
Character string of team ID
}
\description{
Find team ID for year
}
\examples{
find_team_id("Nebraska", 2024)
}
