% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_teams.R
\name{get_teams}
\alias{get_teams}
\title{Get all team IDs for a particular year}
\usage{
get_teams(year, division = 1, sport = "WVB")
}
\arguments{
\item{year}{Year for fall volleyball season.}

\item{division}{NCAA division (must be 1, 2, or 3 for women's and 1 or 3
for men's).}

\item{sport}{Three letter abbreviation for NCAA sport (must be upper case;
for example "WVB" for women's volleyball and "MVB" for men's volleyball).}
}
\value{
Data frame of all teams, their team ID, division, conference, and year.
}
\description{
Get all team IDs for a particular year
}
\examples{
\dontrun{
get_teams(2024)
get_teams(2023, division = 3, sport = "MVB")
}
}
