% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_stats.R
\name{group_player_stats}
\alias{group_player_stats}
\title{Aggregate player statistics}
\usage{
group_player_stats(...)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{group_player_stats()} was renamed to \code{group_stats()} to allow an expanded
functionality.
}
\keyword{internal}
