% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_pbp.R
\name{match_pbp}
\alias{match_pbp}
\title{Extract play-by-play information for a particular match}
\usage{
match_pbp(contest = NULL)
}
\arguments{
\item{contest}{Contest ID determined by NCAA for match. To find ID, use
\code{\link[=find_team_contests]{find_team_contests()}} for a team and season.}
}
\value{
Returns a data frame of set number, teams, score, event, and player responsible
for the event.
}
\description{
The NCAA's page for a match/contest includes a tab called
"Play By Play". This function extracts the tables of play-by-play information
for each set.
}
\examples{
\donttest{
match_pbp(contest = "6080706")
}
}
