% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mvb_teams}
\alias{mvb_teams}
\title{NCAA Men's Volleyball Teams 2020-2024}
\format{
A data frame with 873 rows and 6 columns:
\describe{
\item{team_id}{Team ID for season/year}
\item{team_name}{Team name}
\item{conference_id}{Conference ID}
\item{conference}{Conference name}
\item{div}{NCAA division number (1 or 3)}
\item{yr}{Year for fall of season}
}
}
\source{
\url{https://stats.ncaa.org}
}
\usage{
mvb_teams
}
\description{
This data frame includes all men's NCAA Division 1 and 3 teams from
2020-2024.
}
\examples{
head(mvb_teams)
}
\seealso{
Other data sets: 
\code{\link{ncaa_conferences}},
\code{\link{ncaa_sports}},
\code{\link{ncaa_teams}},
\code{\link{wvb_teams}}
}
\concept{data sets}
\keyword{datasets}
