% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{request_live_url}
\alias{request_live_url}
\title{Submit URL request via live browser}
\usage{
request_live_url(url)
}
\arguments{
\item{url}{URL for request.}
}
\description{
Submit URL request via live browser
}
\note{
This function \strong{requires internet connectivity} as it checks the
\href{https://stats.ncaa.org}{NCAA website} for information.
}
\keyword{internal}
