% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{request_url}
\alias{request_url}
\title{Submit URL request, check, and return response}
\usage{
request_url(url, timeout = 5)
}
\arguments{
\item{url}{URL for request.}

\item{timeout}{Numeric of maximum number of seconds to wait for timeout.}
}
\description{
Submit URL request, check, and return response
}
\keyword{internal}
