% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_player_stats.R
\name{team_player_stats}
\alias{team_player_stats}
\title{Extract player season statistics}
\usage{
team_player_stats(team_id)
}
\arguments{
\item{team_id}{Team ID determined by NCAA for season. To find ID, use
`find_team_id()`.}
}
\value{
Tibble of player statistics. Note that hometown and high school were added
in 2024.
}
\description{
Extract player season statistics
}
\examples{
\dontrun{
team_player_stats("585290")
team_player_stats(find_team_id("Nebraska", 2023))
team_player_stats(find_team_id("UCLA", 2023, sport = "MVB"))
}
}
