% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_match_stats.R, R/team_season_stats.R
\name{team_season_stats}
\alias{team_season_stats}
\title{Extract teams statistics for all matches in a particular season}
\usage{
team_season_stats(team = NULL, opponent = FALSE, sport = "WVB")

team_season_stats(team = NULL, opponent = FALSE, sport = "WVB")
}
\arguments{
\item{team}{Name of school. Must match name used by NCAA. Find exact team
name with \code{\link[=find_team_name]{find_team_name()}}.}

\item{opponent}{Logical indicating whether to include team's stats
(FALSE) or opponent's stats (TRUE). Default is set to FALSE, returning
team stats.}

\item{sport}{Three letter abbreviation for NCAA sport (must be upper case;
for example "WVB" for women's volleyball and "MVB" for men's volleyball).}
}
\value{
Returns a list that includes arena, coach, schedule, and record information.

Returns a data frame of summary team statistics for each season.
}
\description{
The NCAA's main page for a team includes a tab called "Schedule/Results".
This function extracts information about the team's venue, coach, and
records, as well as the table of the schedule and results. This returns a
list, so you can subset specific components with \code{$} (e.g., for coach
information from an object called \code{output}, use \code{output$coach}).

The NCAA's main page for a team includes a tab called "Game By Game"
and a section called "Career Totals". Though the page only shows one
season's worth of information, we can extract season summary stats
starting with 2001. We have included the conference starting with 2020
(conference data for previous seasons is not currently available).
}
\examples{
\dontrun{
team_season_stats(team_id = "585290")
team_season_stats(team_id = find_team_id("Nebraska", 2024))
team_season_stats(team_id = find_team_id("UCLA", 2023, sport = "MVB"))
}
\dontrun{
team_season_stats(team = "Nebraska")
team_season_stats(team = "Nebraska", opponent = TRUE)
team_season_stats(team = "UCLA", sport = "MVB")
}
}
\seealso{
Other functions that extract team statistics: 
\code{\link{player_season_stats}()},
\code{\link{team_season_info}()}

Other functions that extract team statistics: 
\code{\link{player_season_stats}()},
\code{\link{team_season_info}()}
}
\concept{functions that extract team statistics}
