% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wvb_teams}
\alias{wvb_teams}
\title{NCAA Women's Volleyball Teams 2020-2024}
\format{
A data frame with 5,289 rows and 6 columns:
\describe{
\item{team_id}{Team ID for season/year}
\item{team_name}{Team name}
\item{conference_id}{Conference ID}
\item{conference}{Conference name}
\item{div}{NCAA division number (1, 2, or 3)}
\item{yr}{Year for fall of season}
}
}
\source{
\url{https://stats.ncaa.org}
}
\usage{
wvb_teams
}
\description{
This data frame includes all women's NCAA Division 1, 2, and 3 teams from
2020-2024.
}
\examples{
head(wvb_teams)
}
\seealso{
Other data sets: 
\code{\link{mvb_teams}},
\code{\link{ncaa_conferences}},
\code{\link{ncaa_sports}},
\code{\link{ncaa_teams}}
}
\concept{data sets}
\keyword{datasets}
